refactor: migrate service implementations into the counter crate
