--- router/src/adapters/mod.rs
@@ -1,1 +1,0 @@
-pub mod counter_handler_impl;
