+++ router/src/creator/queue.rs
@@ -1,0 +1,120 @@
+use std::sync::{Arc, Mutex};
+use std::time::{Duration, Instant};
+
+use tracing::{error, warn};
+
+use crate::ingress::types::TaskRequest;
+
+pub trait TaskQueue: Send + Sync {
+    #[allow(dead_code)]
+    fn push(&self, task: TaskRequest);
+
+    fn pop(&self) -> Option<TaskRequest>;
+}
+
+#[derive(Clone)]
+pub struct SimpleTaskQueue {
+    queue: Arc<Mutex<Vec<TaskRequest>>>,
+    timeout_ms: u64,
+    max_retries: u32,
+}
+
+impl SimpleTaskQueue {
+    pub fn new() -> Self {
+        Self {
+            queue: Arc::new(Mutex::new(Vec::new())),
+            timeout_ms: 1000,
+            max_retries: 3,
+        }
+    }
+
+    #[allow(dead_code)]
+    pub fn with_timeout(timeout_ms: u64) -> Self {
+        Self {
+            queue: Arc::new(Mutex::new(Vec::new())),
+            timeout_ms,
+            max_retries: 3,
+        }
+    }
+
+    #[allow(dead_code)]
+    pub fn with_config(timeout_ms: u64, max_retries: u32) -> Self {
+        Self {
+            queue: Arc::new(Mutex::new(Vec::new())),
+            timeout_ms,
+            max_retries,
+        }
+    }
+
+    fn try_lock_with_timeout(&self) -> Result<std::sync::MutexGuard<'_, Vec<TaskRequest>>, String> {
+        let start_time = Instant::now();
+        let timeout_duration = Duration::from_millis(self.timeout_ms);
+
+        for attempt in 0..self.max_retries {
+            match self.queue.try_lock() {
+                Ok(guard) => return Ok(guard),
+                Err(_) => {
+                    if start_time.elapsed() >= timeout_duration {
+                        return Err(format!(
+                            "Failed to acquire lock after {}ms timeout ({} attempts)",
+                            self.timeout_ms,
+                            attempt + 1
+                        ));
+                    }
+
+                    std::thread::sleep(Duration::from_millis(1));
+                }
+            }
+        }
+
+        Err(format!(
+            "Failed to acquire lock after {} retries",
+            self.max_retries
+        ))
+    }
+}
+
+impl Default for SimpleTaskQueue {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl TaskQueue for SimpleTaskQueue {
+    fn push(&self, task: TaskRequest) {
+        match self.try_lock_with_timeout() {
+            Ok(mut queue) => {
+                queue.push(task);
+            }
+            Err(e) => {
+                error!("Failed to push task to queue: {}", e);
+                warn!("Task dropped due to lock timeout: {:?}", task);
+            }
+        }
+    }
+
+    fn pop(&self) -> Option<TaskRequest> {
+        match self.try_lock_with_timeout() {
+            Ok(mut queue) => queue.pop(),
+            Err(e) => {
+                error!("Failed to pop task from queue: {}", e);
+                None
+            }
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct CreatorConfig {
+    pub polling_interval_ms: u64,
+    pub timeout_ms: u64,
+}
+
+impl Default for CreatorConfig {
+    fn default() -> Self {
+        Self {
+            polling_interval_ms: 100,
+            timeout_ms: 5000,
+        }
+    }
+}
