+++ router/src/ingress/types.rs
@@ -1,1 +1,18 @@
-pub use commonware_usecase_counter::creator::{TaskRequest, TaskResponse};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct TaskRequestBody {
+    pub metadata: HashMap<String, String>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct TaskRequest {
+    pub body: TaskRequestBody,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct TaskResponse {
+    pub success: bool,
+    pub message: String,
+}
