+++ router/src/orchestrator/tests/generic.rs
@@ -1,9 +1,9 @@
+use super::task_data::TestTaskData;
 use crate::creator::MockCreator;
 use crate::creator::core::Creator;
 use crate::executor::MockExecutor;
 use crate::orchestrator::generic::{Orchestrator, OrchestratorConfig};
 use commonware_avs_core::validator::MockValidator;
-use commonware_usecase_counter::creator::CounterTaskData;
 use std::time::Duration;
 
 use super::helpers::{contributor, signer};
@@ -22,7 +22,7 @@ async fn test_orchestrator_new() {
         threshold: 2,
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new();
+    let task_creator = MockCreator::<TestTaskData>::new();
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
 
@@ -54,7 +54,7 @@ async fn test_orchestrator_task_creator_metadata() {
     let signer = signer::create_test_signer();
     let (contributors, g1_map) = contributor::create_test_contributors();
 
-    let metadata = CounterTaskData {
+    let metadata = TestTaskData {
         var1: "custom_key".to_string(),
         var2: "custom_value".to_string(),
         var3: "42".to_string(),
@@ -67,7 +67,7 @@ async fn test_orchestrator_task_creator_metadata() {
         threshold: 2,
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new().with_metadata(metadata.clone());
+    let task_creator = MockCreator::<TestTaskData>::new().with_metadata(metadata.clone());
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
 
@@ -98,7 +98,7 @@ async fn test_orchestrator_executor_access() {
         threshold: 2,
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new();
+    let task_creator = MockCreator::<TestTaskData>::new();
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
 
@@ -129,7 +129,7 @@ async fn test_orchestrator_validator_access() {
         threshold: 2,
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new();
+    let task_creator = MockCreator::<TestTaskData>::new();
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
 
@@ -160,7 +160,7 @@ async fn test_orchestrator_config_creation() {
         threshold: 3,
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new();
+    let task_creator = MockCreator::<TestTaskData>::new();
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
 
@@ -200,7 +200,7 @@ async fn test_orchestrator_threshold_validation() {
         threshold: 3, // Equal to number of contributors
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new();
+    let task_creator = MockCreator::<TestTaskData>::new();
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
 
@@ -236,7 +236,7 @@ async fn test_orchestrator_component_interaction() {
         threshold: 2,
     };
 
-    let task_creator = MockCreator::<CounterTaskData>::new();
+    let task_creator = MockCreator::<TestTaskData>::new();
     let executor = MockExecutor::new();
     let validator = MockValidator::new_success(1);
