+++ usecase-counter/src/router/executor.rs
@@ -1,12 +1,24 @@
-use crate::executor::bls::{BlsSignatureVerificationHandler, convert_non_signer_data};
+use crate::types::CounterTaskData;
 use alloy_primitives::{Bytes, FixedBytes};
 use anyhow::Result;
 use async_trait::async_trait;
+use commonware_avs_bindings::WalletProvider;
 use commonware_avs_bindings::blssigcheckoperatorstateretriever::BLSSigCheckOperatorStateRetriever::getNonSignerStakesAndSignatureReturn;
-use commonware_avs_bindings::counter::{self};
+use commonware_avs_bindings::counter::{self, Counter};
 use commonware_avs_core::traits::ExecutionResult;
-use commonware_usecase_counter::creator::CounterTaskData;
-use commonware_usecase_counter::executor::CounterHandler;
+use commonware_avs_router::executor::bls::{
+    BlsSignatureVerificationHandler, convert_non_signer_data,
+};
+
+pub struct CounterHandler {
+    pub counter: Counter::CounterInstance<(), WalletProvider>,
+}
+
+impl CounterHandler {
+    pub fn new(counter: Counter::CounterInstance<(), WalletProvider>) -> Self {
+        Self { counter }
+    }
+}
 
 #[async_trait]
 impl BlsSignatureVerificationHandler for CounterHandler {
